requirements.txt
langchain>=0.2.16
langchain-openai>=0.1.8
openai>=1.51.0
pandas>=2.2.2
openpyxl>=3.1.2
python-dotenv>=1.0.1
requests>=2.32.3
